Thunks: Rename new-style GL thunks to FGL
