+++ Data/ThunksDB.json
@@ -1,7 +1,7 @@
 {
   "DB": {
     "FGL": {
-      "Library": "libGL-guest.so",
+      "Library": "libFGL-guest.so",
       "Overlay": [
         "/usr/lib/x86_64-linux-gnu/libGL.so",
         "/usr/lib/x86_64-linux-gnu/libGL.so.1",
@@ -36,7 +36,7 @@
         "/lib/x86_64-linux-gnu/libEGL.so.1",
         "/lib/x86_64-linux-gnu/libEGL.so.1.1.0",
 
-	      "/usr/lib/x86_64-linux-gnu/libGLESv1_CM.so",
+	"/usr/lib/x86_64-linux-gnu/libGLESv1_CM.so",
         "/usr/lib/x86_64-linux-gnu/libGLESv1_CM.so.1",
         "/usr/lib/x86_64-linux-gnu/libGLESv1_CM.so.1.2.0",
         "/usr/local/lib/x86_64-linux-gnu/libGLESv1_CM.so",
