+++ ThunkLibs/GuestLibs/CMakeLists.txt
@@ -117,9 +117,9 @@ endfunction()
 #generate(libEGL ${CMAKE_CURRENT_SOURCE_DIR}/../libEGL/libEGL_interface.cpp thunks function_packs function_packs_public)
 #add_guest_lib(EGL)
 
-generate(libGL ${CMAKE_CURRENT_SOURCE_DIR}/../libGL/libGL_interface.cpp thunks function_packs function_packs_public)
-add_guest_lib(GL)
-target_link_libraries(GL-guest PRIVATE X11)
+generate(libFGL ${CMAKE_CURRENT_SOURCE_DIR}/../libFGL/libFGL_interface.cpp thunks function_packs function_packs_public)
+add_guest_lib(FGL)
+target_link_libraries(FGL-guest PRIVATE X11)
 
 # disabled for now, headers are platform specific
 # find_package(SDL2 REQUIRED)
